perf(uint8): stop the chunked wrapper pessimising its first chunk, and cheapen the IP epilogue
Two findings from benchmark runs on an Ice Lake-SP host and an ARM host, both
measured rather than inferred.

First, the chunked wrapper was 8-9.5% slower per element than the plain kernel on
the stretch its first chunk covers. Cause: the first chunk's length was a
compile-time constant, so its loop had a compile-time trip count, and GCC then
split that loop's accumulator and copied it in and out every 64 elements. The
inner product loop went 12 instructions with no register moves to 13 with two;
L2 went 14/0 to 15/2. The later-chunks loop was always fine.

Fixed by two changes to the chunked wrappers only, leaving the plain path alone:

  - the first chunk's length is now a runtime min against the dimension, so the
    trip count is not a constant. Both loops are back to 12/0 and 14/0, matching
    the plain kernels exactly, at every residual. The min also makes the chunked
    wrapper correct at any dimension rather than only past the chunk size.
  - the residual-0 chunks now go through one out-of-line copy of the kernel
    instead of being inlined into all 64 chunked wrappers. One call per 65,536
    elements is unmeasurable, and it drops the AVX512 inner product family from
    41,267 to 34,229 bytes of text, against 14,735 for the plain family alone.

Second, the inner product epilogue. Converting the unsigned total to float and
subtracting in float costs one instruction more than main's integer subtract plus
a single signed convert, and rounds twice instead of once; it measured about 1%
slower across all four IP benchmark groups. Restored the integer form, which is
also what INT8_InnerProduct already does. The cast to int64_t is required because
ret_t is unsigned for uint8, so 1 minus the total would otherwise wrap. Applied
to the scalar kernel too, so scalar and SIMD stay bit-identical, which the
exactness tests assert.

+++ src/VecSim/spaces/IP/IP.cpp
@@ -277,7 +277,10 @@ float INT8_Cosine(const void *pVect1v, const void *pVect2v, size_t dimension) {
 float UINT8_InnerProduct(const void *pVect1v, const void *pVect2v, size_t dimension) {
     const auto *pVect1 = static_cast<const uint8_t *>(pVect1v);
     const auto *pVect2 = static_cast<const uint8_t *>(pVect2v);
-    return 1.0f - static_cast<float>(INTEGER_InnerProductImp(pVect1, pVect2, dimension));
+    // Integer subtract then a single conversion, matching INT8_InnerProduct above. The cast to a
+    // signed type is required because ret_t is unsigned for uint8: 1 - an unsigned total would
+    // wrap. The SIMD kernels do the same, so their results stay bit-identical to this one.
+    return 1 - static_cast<int64_t>(INTEGER_InnerProductImp(pVect1, pVect2, dimension));
 }
 
 float UINT8_Cosine(const void *pVect1v, const void *pVect2v, size_t dimension) {
